LVST enteric costs
This update fixes a bug in the application and calculation of livestock enteric fermentation costs. The bug was that the function cb_lvst_enteric getting called once for each livestock type, but it also itself calculated costs for all livestock types, meaning the costs were getting multiplied by the number of livestock. That is fixed by making the csv instructions search only one livestock type.